+++ symboard/states.py
@@ -17,30 +17,25 @@ from yaml import safe_load
 
 # Imports from this package.
 from symboard.actions import State
-
-
-states_dir = 'symboard/states'
-""" The relative path to the directory containing definitions of states. These
-states can be pre-defined by the Symboard project, or added by users.
-"""
+from settings import STATE_ATTRIBUTE_PRECEDENCE, STATES_DIR
 
 
 def load_yaml():
-    """ Loads all yaml files which can be found in the folder <states_dir>, and
+    """ Loads all yaml files which can be found in the folder <STATES_DIR>, and
     adds them to an object «states», which can then be imported and used
     throughout the project.
     """
     logging.info(
-        f'Creating yaml states object using files from folder «{states_dir}».'
+        f'Creating yaml states object using files from folder «{STATES_DIR}».'
     )
 
     states: Dict[State] = {}
 
-    for _, _, file_names in walk(states_dir):
+    for _, _, file_names in walk(STATES_DIR):
         for file_name in file_names:
             logging.info(f'Importing yaml states from file {file_name}.')
 
-            file_path = states_dir + '/' + file_name
+            file_path = STATES_DIR + '/' + file_name
 
             with open(file_path, 'r') as file_:
                 contents = safe_load(file_)
@@ -48,13 +43,17 @@ def load_yaml():
             for name, attribs in contents.items():
                 new_state = State(
                     name=name,
-                    terminator=attribs['terminator'],
-                ).with_lower(
-                    attribs['lower']
+                    # Using .get allows us to pass in a None argument without
+                    # raising an error.
+                    terminator=attribs.get('terminator'),
                 )
 
-                if attribs.get('upper'):
-                    new_state = new_state.with_upper(attribs['upper'])
+                for attrib in STATE_ATTRIBUTE_PRECEDENCE:
+                    if attribs.get(attrib):
+                        method = new_state.builder_method_from_attrib_name(
+                            attrib
+                        )
+                        new_state = method(attribs[attrib])
 
                 states[name] = new_state
 
@@ -62,7 +61,7 @@ def load_yaml():
 
 
 states = load_yaml()
-""" An object containing all states found inside <states_dir>, which can be
+""" An object containing all states found inside <STATES_DIR>, which can be
 imported and used throughout the project.
 """